Refactored user interface code out of main file
Moved all main code into ui.py.
